+++ lib/app.js
@@ -2,7 +2,7 @@ var request = require('request'),
     config = require('./config'),
     path = require('path'),
     cradle = require('cradle'),    
-    lib = require('../lib'),    
+    lib = require('./lib'),    
     spawn = require('child_process').spawn;
