+++ lib/domains.js
@@ -1,6 +1,6 @@
 var request = require('request'),
     config = require('./config'),
-    lib = require('../lib');
+    lib = require('./lib');
 
 module.exports = {
     post: function(req, res, next) {
