+++ lib/lib.js
